Applied suggestions to the PR
